Silently fail on 'sites' directory not existing.
When determining public file paths if the parent sites directory does not exist
and exception is thrown which can break package command execution as well as
initial installation (assuming config is defined prior to install).

If the sites directory does not exist it is a safe assumption that the files
directories do not exist either so silently failing and returning an empty array
can be used safely and should not cause any other issues.

Catch the base exception as Symfony finder can throw varying exceptions based
on (assume) package versions. InvalidArgumentException as well as
DirectoryNotFound exceptions.